requirements.txt
alabaster==1.0.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.1
babel==2.17.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
docutils==0.21.2
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.16
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pipx==1.7.1
platformdirs==4.3.6
psycopg2==2.9.10
pyasn1==0.4.8
pydantic==2.11.1
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis>=5.0.0
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.0
roman-numerals-py==3.1.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.40
starlette==0.46.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
ujson==5.10.0
urllib3==2.3.0
userpath==1.9.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
wrapt==1.17.2
